user query testing
