+++ server/package.json
@@ -20,6 +20,7 @@
     "express": "^4.17.1",
     "express-graphql": "^0.12.0",
     "graphql": "^15.5.1",
+    "mongodb": "^4.0.0",
     "ts-node": "^10.1.0",
     "typeorm": "^0.2.34",
     "typescript": "^4.3.5"
